Improve performance of lox-rs by avoiding clone operations.
This improves the performance significantly by avoiding as many clone
operations as possible. This is largely done by making things into RCs
or references wherever possible. There is still room for improvement to
get rid of some of the clones here, but performance is improved greatly:
- fib-rec.lox took 4.5 seconds on my machine, now takes 2 seconds.
- fib.lox took 6.5 seconds on my machine, now takes 0.5 seconds.

I updated the programs as well to be more useful for benchmarking, but
maybe those should be new copies so that the small test versions can
still live on their own.